refactor(Bpu): Separate history-related logic from Bpu
